Updated Newton fractal to work with doubles and increased origin limits, but is still bugged.

+++ Fractal Z/shaders/newton/newton_compute.glsl
@@ -1,29 +1,29 @@
 uniform int maxIterations;
 
 // Multiply two complex numbers
-vec2 complexMultiply(vec2 a, vec2 b)
+dvec2 complexMultiply(dvec2 a, dvec2 b)
 {
-	float r = (a.x * b.x) - (a.y * b.y);	// Compute real part
-	float i = (a.x * b.y) + (a.y * b.x);	// Cmopute imaginary part
+	double r = (a.x * b.x) - (a.y * b.y);	// Compute real part
+	double i = (a.x * b.y) + (a.y * b.x);	// Cmopute imaginary part
 
-	return vec2(r, i);
+	return dvec2(r, i);
 }
 
 // Divide two complex numbers (rationalise denominator, divide)
-vec2 complexDivide(vec2 a, vec2 b)
+dvec2 complexDivide(dvec2 a, dvec2 b)
 {
-	vec2 rat = vec2(b.x, -b.y);							// Conjugate of denominator
-	vec2 denominator = complexMultiply(rat, b);			// Multiply denominator by conjugate
-	float divisor = denominator.x;
+	dvec2 rat = dvec2(b.x, -b.y);							// Conjugate of denominator
+	dvec2 denominator = complexMultiply(rat, b);			// Multiply denominator by conjugate
+	double divisor = denominator.x;
 	a = complexMultiply(a, rat);
-	vec2 result = vec2(a.x / divisor, a.y / divisor);	// Divide
+	dvec2 result = dvec2(a.x / divisor, a.y / divisor);		// Divide
 
 	return result;
 }
 
 // This is the function being used
 // f(z) = z^4 - 1
-vec2 calcFx(vec2 a)
+dvec2 calcFx(dvec2 a)
 {
 	a = complexMultiply(a, a);	// z^2
 	a = complexMultiply(a, a);	// z^2^2 = z^4
@@ -34,9 +34,9 @@ vec2 calcFx(vec2 a)
 
 // This is the derivative of the function being used
 // f'(z) = 4z^3
-vec2 calcFpx(vec2 a)
+dvec2 calcFpx(dvec2 a)
 {
-	vec2 old = a;					// Save old z
+	dvec2 old = a;					// Save old z
 	a = complexMultiply(a, a);		// z^2
 	a = complexMultiply(a, old);	// z^2 * z = z^3
 	a = 4 * a;						// z^3 * 4
@@ -45,28 +45,28 @@ vec2 calcFpx(vec2 a)
 }
 
 // Compute
-float compute(vec2 pos)
+float compute(dvec2 pos)
 {
 	// Roots of f(z)
-	vec2 roots[4];
-	roots[0] = vec2(1.0f, 0.0f);	// 1
-	roots[1] = vec2(-1.0f, 0.0f);	// -1
-	roots[2] = vec2(0.0f, 1.0f);	// i
-	roots[3] = vec2(0.0f, -1.0f);	// -i
+	dvec2 roots[4];
+	roots[0] = dvec2(1.0f, 0.0f);	// 1
+	roots[1] = dvec2(-1.0f, 0.0f);	// -1
+	roots[2] = dvec2(0.0f, 1.0f);	// i
+	roots[3] = dvec2(0.0f, -1.0f);	// -i
 
 	// Initiaise variables
-	vec2 c = pos;
-	vec2 fx = vec2(0.0f, 0.0f);
-	vec2 fpx = vec2(0.0f, 0.0f);
+	dvec2 c = pos;
+	dvec2 fx = vec2(0.0f, 0.0f);
+	dvec2 fpx = vec2(0.0f, 0.0f);
 
-	vec2 norm = vec2(0.0f, 0.0f);
+	dvec2 norm = vec2(0.0f, 0.0f);
 	int root = -1;
-	float tolerance = 0.001f;
-	float dist = 0.0f;
+	double tolerance = 0.001f;
+	double dist = 0.0f;
 	int iterations = 0;
 
 	// Previous closest to each root, used in smooth colouring
-	float prev[4];
+	double prev[4];
 	prev[0] = tolerance;
 	prev[1] = tolerance;
 	prev[2] = tolerance;
@@ -108,8 +108,8 @@ float compute(vec2 pos)
 
 	if (root != -1)
 	{
-		float logi = (log(tolerance) - log(prev[root])) / (log(dist) - log(prev[root]));
-		shade = ((iterations + logi) / maxIterations) * 1.5f;
+		float logi = (log(float(tolerance)) - log(float(prev[root]))) / (log(float(dist)) - log(float(prev[root])));
+		shade = ((iterations + logi) / float(maxIterations)) * 1.5f;
 		shade = 1.0f - shade;
 	}
